+++ .github/workflows/ci.yml
@@ -50,11 +50,11 @@ jobs:
       - run: go build ./...
       - run: go test ./...
 
-  infra:
+  lambda:
     runs-on: ubuntu-latest
     defaults:
       run:
-        working-directory: ./infra
+        working-directory: ./infra/lambda
     steps:
       - uses: actions/checkout@v6
       - uses: pnpm/action-setup@v4
@@ -64,16 +64,10 @@ jobs:
         with:
           node-version: "24"
           cache: "pnpm"
-          cache-dependency-path: "infra/pnpm-lock.yaml"
-      - uses: hashicorp/setup-terraform@v3
-        with:
-          terraform_version: "1.5.0"
+          cache-dependency-path: "infra/lambda/pnpm-lock.yaml"
       - run: pnpm install --no-frozen-lockfile
-      - run: npm install -g cdktf-cli
+      - run: pnpm type-check
       - run: pnpm build
-      - run: pnpm synth
-      - run: ls -la cdktf.out/stacks/tku/ || echo "Synth output not found"
-      - run: test -f cdktf.out/stacks/tku/cdk.tf.json || (echo "Terraform JSON not generated" && exit 1)
 
   opentofu:
     runs-on: ubuntu-latest
