+++ .tool-versions
@@ -4,9 +4,6 @@
 
 # Frontend and infra use Node.js, backend API uses Go.
 
-# Terraform is retained temporarily while the legacy CDKTF configuration remains.
-
 nodejs 24.16.0 # Node.js LTS version for frontend and infrastructure
 golang 1.26.3 # Latest stable Go version for backend API
-terraform 1.15.5 # Temporary: required by the legacy CDKTF configuration
 opentofu 1.12.0 # IaC for the OpenTofu migration
