--- infra/.prettierrc
@@ -1,8 +1,0 @@
-{
-    printWidth: 150,
-    tabWidth: 4,
-    semi: false,
-    singleQuote: true,
-    arrowParens: 'always',
-    jsxBracketSameLine: true,
-}
