--- infra/libs/convert.ts
@@ -1,3 +1,0 @@
-export function snakeToCamelCase(str: string): string {
-    return str.replace(/-([a-z])/g, (_match, letter) => letter.toUpperCase())
-}
