--- infra/package.json
@@ -1,36 +1,0 @@
-{
-  "name": "infra",
-  "version": "1.0.0",
-  "main": "main.js",
-  "types": "main.ts",
-  "license": "MPL-2.0",
-  "private": true,
-  "packageManager": "pnpm@10.34.1",
-  "scripts": {
-    "get": "cdktf get",
-    "build": "tsc",
-    "synth": "cdktf synth",
-    "compile": "tsc --pretty",
-    "watch": "tsc -w",
-    "test": "jest",
-    "test:watch": "jest --watch",
-    "upgrade": "npm i cdktf@latest cdktf-cli@latest",
-    "upgrade:next": "npm i cdktf@next cdktf-cli@next"
-  },
-  "dependencies": {
-    "@cdktf/provider-aws": "21.22.1",
-    "cdktf": "^0.21.0",
-    "constructs": "^10.4.2",
-    "dotenv": "^17.2.1",
-    "esbuild": "^0.28.0",
-    "node-fetch": "^3.3.2"
-  },
-  "devDependencies": {
-    "@types/jest": "^30.0.0",
-    "@types/node": "^24.3.0",
-    "jest": "^30.1.0",
-    "ts-jest": "^29.4.1",
-    "ts-node": "^10.9.2",
-    "typescript": "^5.9.2"
-  }
-}
