--- infra/resources/ec2/userData.ts
@@ -1,3 +1,0 @@
-// ユーザーデータを取得する
-export const getUserData = (clusterName: string) => `#!/bin/bash
-echo ECS_CLUSTER=${clusterName} >> /etc/ecs/ecs.config;`
