--- infra/resources/ecs/tasks/api.json
@@ -1,126 +1,0 @@
-{
-    "ipcMode": null,
-    "executionRoleArn": "arn:aws:iam::418549683327:role/tku-ecs-task-exec-role",
-    "containerDefinitions": [
-        {
-            "image": "418549683327.dkr.ecr.ap-northeast-1.amazonaws.com/tku_api",
-            "dnsSearchDomains": null,
-            "environmentFiles": null,
-            "logConfiguration": null,
-            "entryPoint": [],
-            "portMappings": [
-                {
-                    "hostPort": 8080,
-                    "protocol": "tcp",
-                    "containerPort": 8080
-                }
-            ],
-            "command": [
-                "sh",
-                "./docker/api/script/prod/command.sh"
-            ],
-            "linuxParameters": null,
-            "cpu": 0,
-            "environment": [],
-            "resourceRequirements": null,
-            "ulimits": null,
-            "dnsServers": null,
-            "mountPoints": [],
-            "workingDirectory": null,
-            "secrets": [
-                {
-                    "valueFrom": "arn:aws:secretsmanager:ap-northeast-1:418549683327:secret:tku-secrets-manager-wndNhr:CLIENT_URL::",
-                    "name": "CLIENT_URL"
-                },
-                {
-                    "valueFrom": "arn:aws:secretsmanager:ap-northeast-1:418549683327:secret:tku-secrets-manager-wndNhr:API_BASE_URL::",
-                    "name": "API_BASE_URL"
-                },
-                {
-                    "valueFrom": "arn:aws:secretsmanager:ap-northeast-1:418549683327:secret:tku-secrets-manager-wndNhr:CREATOR_NAME::",
-                    "name": "CREATOR_NAME"
-                },
-                {
-                    "valueFrom": "arn:aws:secretsmanager:ap-northeast-1:418549683327:secret:tku-secrets-manager-wndNhr:DB_NAME::",
-                    "name": "DB_NAME"
-                },
-                {
-                    "valueFrom": "arn:aws:secretsmanager:ap-northeast-1:418549683327:secret:tku-secrets-manager-wndNhr:DB_PASS::",
-                    "name": "DB_PASS"
-                },
-                {
-                    "valueFrom": "arn:aws:secretsmanager:ap-northeast-1:418549683327:secret:tku-secrets-manager-wndNhr:DB_USER::",
-                    "name": "DB_USER"
-                },
-                {
-                    "valueFrom": "arn:aws:secretsmanager:ap-northeast-1:418549683327:secret:tku-secrets-manager-wndNhr:ENV::",
-                    "name": "ENV"
-                },
-                {
-                    "valueFrom": "arn:aws:secretsmanager:ap-northeast-1:418549683327:secret:tku-secrets-manager-wndNhr:MYSQL_HOST::",
-                    "name": "MYSQL_HOST"
-                },
-                {
-                    "valueFrom": "arn:aws:secretsmanager:ap-northeast-1:418549683327:secret:tku-secrets-manager-wndNhr:MYSQL_ROOT_PASSWORD::",
-                    "name": "MYSQL_ROOT_PASSWORD"
-                },
-                {
-                    "valueFrom": "arn:aws:secretsmanager:ap-northeast-1:418549683327:secret:tku-secrets-manager-wndNhr:AWS_SECRET_ACCESS_KEY::",
-                    "name": "AWS_SECRET_ACCESS_KEY"
-                },
-                {
-                    "valueFrom": "arn:aws:secretsmanager:ap-northeast-1:418549683327:secret:tku-secrets-manager-wndNhr:AWS_ACCESS_KEY_ID::",
-                    "name": "AWS_ACCESS_KEY_ID"
-                },
-                {
-                    "valueFrom": "arn:aws:secretsmanager:ap-northeast-1:418549683327:secret:tku-secrets-manager-wndNhr:AWS_REGION::",
-                    "name": "AWS_REGION"
-                },
-                {
-                    "valueFrom": "arn:aws:secretsmanager:ap-northeast-1:418549683327:secret:tku-secrets-manager-wndNhr:API_BUCKET_NAME::",
-                    "name": "API_BUCKET_NAME"
-                },
-                {
-                    "valueFrom": "arn:aws:secretsmanager:ap-northeast-1:418549683327:secret:tku-secrets-manager-wndNhr:SEND_GRID_API_KEY::",
-                    "name": "SEND_GRID_API_KEY"
-                }
-            ],
-            "dockerSecurityOptions": null,
-            "memory": 1024,
-            "memoryReservation": null,
-            "volumesFrom": [],
-            "stopTimeout": null,
-            "startTimeout": null,
-            "firelensConfiguration": null,
-            "dependsOn": null,
-            "disableNetworking": null,
-            "interactive": null,
-            "healthCheck": null,
-            "essential": true,
-            "links": null,
-            "hostname": null,
-            "extraHosts": null,
-            "pseudoTerminal": null,
-            "user": null,
-            "readonlyRootFilesystem": null,
-            "dockerLabels": null,
-            "systemControls": null,
-            "privileged": null,
-            "name": "api"
-        }
-    ],
-    "placementConstraints": [],
-    "memory": null,
-    "taskRoleArn": "arn:aws:iam::418549683327:role/tku-ecs-task-exec-role",
-    "family": "tku-api",
-    "pidMode": null,
-    "requiresCompatibilities": [
-        "EC2"
-    ],
-    "networkMode": null,
-    "runtimePlatform": null,
-    "cpu": null,
-    "inferenceAccelerators": null,
-    "proxyConfiguration": null,
-    "volumes": []
-}
