--- infra/resources/ecs/tasks/https.json
@@ -1,38 +1,0 @@
-{
-    "containerDefinitions": [
-        {
-            "name": "https",
-            "image": "steveltn/https-portal",
-            "cpu": 0,
-            "memory": 128,
-            "portMappings": [
-                {
-                    "containerPort": 80,
-                    "hostPort": 80,
-                    "protocol": "tcp"
-                },
-                {
-                    "containerPort": 443,
-                    "hostPort": 443,
-                    "protocol": "tcp"
-                }
-            ],
-            "essential": true,
-            "entryPoint": [],
-            "command": [],
-            "environment": [],
-            "mountPoints": [],
-            "volumesFrom": [],
-            "secrets": [
-                {
-                    "name": "DOMAINS",
-                    "valueFrom": "arn:aws:secretsmanager:ap-northeast-1:418549683327:secret:tku-secrets-manager-wndNhr:DOMAINS::"
-                },
-                {
-                    "name": "STAGE",
-                    "valueFrom": "arn:aws:secretsmanager:ap-northeast-1:418549683327:secret:tku-secrets-manager-wndNhr:STAGE::"
-                }
-            ]
-        }
-    ]
-}
